Fix CI build issues on Ubuntu and macOS
- Create header symlinks at configure time instead of build time
  This ensures all symlinks exist before any compilation starts
- Fix Qt6 implicit conversion warning in guilogger
  Cast qsizetype to int explicitly for getNumMultiChannels()
- Removes race condition that caused 'selforg/storeable.h not found' error

+++ guilogger/src/channeldata.h
@@ -120,7 +120,7 @@ public:
   const QVector<ChannelVals>& getData() const { return data; }
   const QVector<ChannelInfo>& getInfos() const { return channels; }
   int getNumChannels() const { return numchannels; }
-  int getNumMultiChannels() const { return multichannels.size(); }
+  int getNumMultiChannels() const { return static_cast<int>(multichannels.size()); }
     const MultiChannels& getMultiChannels() const { return multichannels; }
   int getTime() const { return time; }
